+++ addons/ai_assistant/models/__init__.py
@@ -2,4 +2,5 @@ from . import ai_audit_log
 from . import ai_tool_definition
 from . import ai_prompt_template
 from . import ai_document_chunk
+from . import ai_conversation
 from . import rag_reindex